producer/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.6
kafka-python==2.0.2